component Logic delegation
